All service moved to SelectCategoryActivity. Service removed from SurakshitActivity and SocialShareActivity
