+++ app/src/main/res/layout/content_main.xml
@@ -11,7 +11,7 @@
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context=".SurakshitActivity"
     tools:showIn="@layout/activity_surakshit">
-
+<!--
     <android.support.v7.widget.CardView
         xmlns:card_view="http://schemas.android.com/apk/res-auto"
         xmlns:android="http://schemas.android.com/apk/res/android"
@@ -152,7 +152,7 @@
         </RelativeLayout>
     </android.support.v7.widget.CardView>
 
-
+-->
 
     <android.support.v7.widget.CardView
         xmlns:card_view="http://schemas.android.com/apk/res-auto"
